game/requirements.txt
requests
django
django-core
djangorestframework
django-cors-headers
psycopg2-binary
PyJWT
